moved modals to modal folder. made some fixes to things for jshint and fixed soemthings
